requirements.txt
django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ipython==8.20.0
bcrypt==4.1.2
pyotp==2.9.0
requests==2.31.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
Faker==23.2.1
drf-spectacular==0.27.1
